+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add test and update README.md
 - Refactor HTML function to use String.raw
+- Add new function `watchEffect`
 
 ## 1.2.6- 2021-01-01